pyproject.toml
[project]
name = "sql-mcp-server"
version = "0.1.0"
description = "a sql mcp server here!"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=2.7.1",
    "fastapi>=0.115.12",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
]
